Revert to using external browser for authentication
Otherwise the TabBarVC will try to display on top of the SFSafariVC and this will fail.
All the code to use the SFSafariVC is still there, just disabled. Re-enabling it, will require bigger changes to the Sample All, or just use a different mechanism other than the NotificationCenter for handling auth callbacks (maybe an Authorization Delegate?)